@@ -1,31 +1,57 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.panteleyev</groupId>
     <artifactId>password-manager</artifactId>
-    <version>19.1.0</version>
+    <version>20.1.0</version>
     <packaging>jar</packaging>
 
     <properties>
         <app.main.class>org.panteleyev.pwdmanager.PasswordManagerApplication</app.main.class>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <gson.version>2.8.5</gson.version>
-        <commons.version>19.1.2</commons.version>
-        <javafx-version>11.0.2</javafx-version>
+        <maven.compiler.source>13</maven.compiler.source>
+        <maven.compiler.target>13</maven.compiler.target>
+        <commons-version>20.1.0</commons-version>
+        <javafx-version>13.0.1</javafx-version>
+        <testng-version>6.14.3</testng-version>
+        <!-- Plugins -->
+        <mavenToolchainPluginVersion>3.0.0</mavenToolchainPluginVersion>
+        <mavenCompilerPluginVersion>3.8.1</mavenCompilerPluginVersion>
+        <mavenJarPluginVersion>3.2.0</mavenJarPluginVersion>
+        <mavenDependencyPluginVersion>3.1.1</mavenDependencyPluginVersion>
     </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>org.openjfx</groupId>
+                <artifactId>javafx-swing</artifactId>
+                <version>${javafx-version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.openjfx</groupId>
+                <artifactId>javafx-web</artifactId>
+                <version>${javafx-version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.openjfx</groupId>
+                <artifactId>javafx-media</artifactId>
+                <version>${javafx-version}</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 
     <dependencies>
         <dependency>
             <groupId>org.panteleyev</groupId>
             <artifactId>org-panteleyev-commons-fx</artifactId>
-            <version>${commons.version}</version>
+            <version>${commons-version}</version>
         </dependency>
         <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>crypto</artifactId>
-            <version>18.1.0</version>
+            <groupId>org.panteleyev</groupId>
+            <artifactId>org-panteleyev-commons-crypto</artifactId>
+            <version>${commons-version}</version>
         </dependency>
         <dependency>
             <groupId>org.openjfx</groupId>
