@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2019, Petr Panteleyev <petr@panteleyev.org>
+ * Copyright (c) 2016, 2020, Petr Panteleyev <petr@panteleyev.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,7 +25,9 @@
  */
 package org.panteleyev.pwdmanager;
 
-import javafx.geometry.Pos;
+import javafx.geometry.HPos;
+import javafx.geometry.Insets;
+import javafx.geometry.VPos;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.Hyperlink;
 import javafx.scene.control.Label;
@@ -36,20 +38,25 @@ import javafx.scene.image.ImageView;
 import javafx.scene.input.Clipboard;
 import javafx.scene.input.ClipboardContent;
 import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.ColumnConstraints;
 import javafx.scene.layout.GridPane;
+import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
 import org.panteleyev.pwdmanager.model.Field;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
-import java.util.ResourceBundle;
+import static org.panteleyev.commons.fx.FXFactory.newLabel;
+import static org.panteleyev.pwdmanager.PasswordManagerApplication.RB;
 
 class CardViewer extends BorderPane implements Styles {
-    private final ResourceBundle rb = PasswordManagerApplication.getBundle();
+    private static final double LEFT_WIDTH = 40.0;
+    private static final double RIGHT_WIDTH = 100.0 - LEFT_WIDTH;
+    private static final String MASK = "*****";
 
     private final GridPane grid = new GridPane();
-    private final Label noteLabel = new Label(rb.getString("label.notesNoSemicolon"),
+    private final Label noteLabel = newLabel(RB, "label.notesNoSemicolon",
         new ImageView(Picture.NOTE.getImage()));
     private final Label noteViewer = new Label();
 
@@ -59,13 +66,25 @@ class CardViewer extends BorderPane implements Styles {
 
     private void initialize() {
         grid.getStyleClass().add(GRID_PANE);
-        grid.setAlignment(Pos.TOP_CENTER);
+
+        var col1 = new ColumnConstraints();
+        col1.setPercentWidth(LEFT_WIDTH);
+        col1.setHgrow(Priority.NEVER);
+        col1.setHalignment(HPos.RIGHT);
+        var col2 = new ColumnConstraints();
+        col2.setPercentWidth(RIGHT_WIDTH);
+        col2.setHgrow(Priority.NEVER);
+        col2.setHalignment(HPos.LEFT);
+        grid.getColumnConstraints().setAll(col1, col2);
+        grid.setFocusTraversable(false);
 
         var vBox = new VBox(
             grid,
             noteLabel,
             noteViewer
         );
+
+        VBox.setMargin(grid, new Insets(10, 5, 10, 5));
 
         noteLabel.getStyleClass().add("noteLabel");
         noteViewer.getStyleClass().add("noteViewer");
@@ -76,14 +95,12 @@ class CardViewer extends BorderPane implements Styles {
         pane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
         pane.setVbarPolicy(ScrollPane.ScrollBarPolicy.AS_NEEDED);
         setCenter(pane);
-
-        BorderPane.setAlignment(pane, Pos.CENTER);
     }
 
     void setData(List<FieldWrapper> items, String note) {
         grid.getChildren().clear();
 
-        int y = 1;
+        int y = 0;
         for (var field : items) {
             var nameLabel = new Label(field.getName());
             nameLabel.getStyleClass().add("fieldName");
@@ -93,8 +110,7 @@ class CardViewer extends BorderPane implements Styles {
                 valueLabel = new Hyperlink(field.getValue());
                 ((Hyperlink) valueLabel).setOnAction(e -> onHyperlinkClick(field.getValue()));
             } else {
-                valueLabel = new Label(field.getType() == FieldType.HIDDEN ?
-                    "***" : field.getValue());
+                valueLabel = new Label(field.getType() == FieldType.HIDDEN ? MASK : field.getValue());
 
                 valueLabel.setOnMouseClicked(event -> {
                     if (event.getClickCount() > 1) {
@@ -104,9 +120,13 @@ class CardViewer extends BorderPane implements Styles {
             }
 
             valueLabel.setContextMenu(createContextMenu(field));
+            valueLabel.setWrapText(true);
 
-            grid.add(nameLabel, 1, y);
-            grid.add(valueLabel, 2, y++);
+            grid.add(nameLabel, 0, y);
+            grid.add(valueLabel, 1, y++);
+
+            GridPane.setValignment(nameLabel, VPos.TOP);
+            GridPane.setValignment(valueLabel, VPos.TOP);
         }
 
         noteViewer.setVisible(!note.isEmpty());
@@ -123,16 +143,16 @@ class CardViewer extends BorderPane implements Styles {
     }
 
     private void onContentViewDoubleClick(FieldWrapper field, Labeled label) {
-        switch (field.getType()) {
-            case HIDDEN:
-                field.toggleShow();
-                label.setText(field.getShow() ? field.getValue() : "***");
-                break;
+        if (field.getType() != FieldType.HIDDEN) {
+            return;
         }
+
+        field.toggleShow();
+        label.setText(field.getShow() ? field.getValue() : MASK);
     }
 
     private ContextMenu createContextMenu(FieldWrapper field) {
-        var copyMenuItem = new MenuItem("Copy " + field.getName());
+        var copyMenuItem = new MenuItem(RB.getString("menu.ctx.copy") + " '" + field.getName() + "'");
         copyMenuItem.setOnAction(x -> onCopy(field));
 
         return new ContextMenu(copyMenuItem);
