@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2019, Petr Panteleyev <petr@panteleyev.org>
+ * Copyright (c) 2016, 2020, Petr Panteleyev <petr@panteleyev.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -34,15 +34,14 @@ import javafx.scene.layout.GridPane;
 import org.controlsfx.validation.ValidationResult;
 import org.panteleyev.commons.fx.BaseDialog;
 import org.panteleyev.pwdmanager.model.Card;
-import java.util.ResourceBundle;
+import static org.panteleyev.commons.fx.FXFactory.newLabel;
+import static org.panteleyev.pwdmanager.PasswordManagerApplication.RB;
 
 abstract class RecordDialog extends BaseDialog<Card> implements Styles {
-    final ResourceBundle               rb = PasswordManagerApplication.getBundle();
-
-    private final TextField            nameEdit = new TextField();
+    private final TextField nameEdit = new TextField();
     private final ComboBox<RecordType> typeList = new ComboBox<>();
-    private final ComboBox<Picture>    pictureList = new ComboBox<>();
-    private final Label                typeLabel = new Label(rb.getString("label.type"));
+    private final ComboBox<Picture> pictureList = new ComboBox<>();
+    private final Label typeLabel = newLabel(RB, "label.type");
 
     RecordDialog() {
         super(MainWindowController.CSS_PATH);
@@ -50,9 +49,9 @@ abstract class RecordDialog extends BaseDialog<Card> implements Styles {
         var grid = new GridPane();
         grid.getStyleClass().add(GRID_PANE);
 
-        grid.addRow(0, new Label(rb.getString("label.Name")), nameEdit);
+        grid.addRow(0, newLabel(RB, "label.Name"), nameEdit);
         grid.addRow(1, typeLabel, typeList);
-        grid.addRow(2, new Label(rb.getString("label.Icon")), pictureList);
+        grid.addRow(2, newLabel(RB, "label.Icon"), pictureList);
 
         nameEdit.setPrefColumnCount(25);
 
@@ -84,7 +83,7 @@ abstract class RecordDialog extends BaseDialog<Card> implements Styles {
 
     void setupValidator() {
         validation.registerValidator(nameEdit, (Control c, String value) ->
-                ValidationResult.fromErrorIf(c, null, nameEdit.getText().isEmpty()));
+            ValidationResult.fromErrorIf(c, null, nameEdit.getText().isEmpty()));
         validation.initInitialDecoration();
     }
 }
